requirements/deployment.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements/deployment.txt --resolver=backtracking requirements/deployment.in
#
asgiref==3.7.2
    # via django
certifi==2023.7.22
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.2.0
    # via requests
cryptography==41.0.3
    # via djangorestframework-simplejwt
django==4.2.4
    # via
    #   -r requirements/deployment.in
    #   django-ninja
    #   djangorestframework
    #   djangorestframework-simplejwt
django-ninja==0.22.2
    # via -r requirements/deployment.in
djangorestframework==3.14.0
    # via djangorestframework-simplejwt
djangorestframework-simplejwt[crypto]==5.3.0
    # via -r requirements/deployment.in
idna==3.4
    # via requests
pycparser==2.21
    # via cffi
pydantic==1.10.12
    # via django-ninja
pyjwt==2.8.0
    # via djangorestframework-simplejwt
pytz==2023.3
    # via djangorestframework
requests==2.31.0
    # via -r requirements/deployment.in
sqlparse==0.4.4
    # via django
typing-extensions==4.7.1
    # via
    #   asgiref
    #   pydantic
urllib3==2.0.4
    # via requests
